Moving around partitions so that we can create a simple "flash from zero" bin
We now have two spiffs paritions that we store the ui data in as well as user data (scripts and analysis items).

And we have moved around stuff in flash so that we can produce some bin's that we can flash from 0 to get a full default load, or just update the firmware and admin ui spiff

Also we can now make a web firmware bin and a web ui bin that can be used to update via the web interface.